(FACT-1128) Alert user if we can't manage job groups
Previously, if Facter on Windows was running in a job group that did
not allow detaching child processes, we would fail with an "Access
Denied" error when trying to attach our child processes to our own job
group.

This now checks early in the execute function for whether we are
running as part of such a job group, and alerts the user with an
appropriate error message.