(maint) Fix cross-compilation from linux->windows
The big change here is a rename of our `system_error` function to
`win32_error`. On modern GCCs, `std::system_error` is now defined and
was causing all sorts of errors to be thrown.

This also downcases a bunch of includes and library names that are
uppercased by convention on Windows, but that are actually all
lowercase filenames. This is needed for windows compilation to work on
a case-sensitive filesystem.